Day 5 Challenge 1 More Practical
My initial solution to challenge 1 would have made the user turn off a light before they could turn on another.  Leaving them to travel in the dark.  
This new solution lets the user travel about the ship with the previous light on and when they flip a new switch in a new room it will take priority.  If you go back to a room with the switch on, just turn it off then on again to regain priority in that room.